# Define namespaces inside generic files