--- a/src/arm64/begin_implementation.h
+++ b/src/arm64/begin_implementation.h
@@ -1,5 +1,9 @@
 #include "simdjson.h"
 #include "arm64/implementation.h"
-#include "arm64/intrinsics.h" // Generally need to be included outside SIMDJSON_TARGET_REGION
+#include "arm64/intrinsics.h"
 
 #define SIMDJSON_IMPLEMENTATION arm64
+
+#include "arm64/bitmanipulation.h"
+#include "arm64/bitmask.h"
+#include "arm64/simd.h"