--- a/src/arm64/bitmask.h
+++ b/src/arm64/bitmask.h
@@ -3,8 +3,6 @@
 
 #include "simdjson.h"
 
-#include "arm64/intrinsics.h"
-
 namespace simdjson {
 namespace arm64 {
 
@@ -38,6 +36,6 @@ really_inline uint64_t prefix_xor(uint64_t bitmask) {
 
 } // namespace arm64
 } // namespace simdjson
-UNTARGET_REGION
+SIMDJSON_UNTARGET_REGION
 
 #endif