--- a/src/arm64/bitmask.h
+++ b/src/arm64/bitmask.h
@@ -1,8 +1,6 @@
 #ifndef SIMDJSON_ARM64_BITMASK_H
 #define SIMDJSON_ARM64_BITMASK_H
 
-#include "simdjson.h"
-
 namespace simdjson {
 namespace arm64 {