--- a/src/arm64/dom_parser_implementation.h
+++ b/src/arm64/dom_parser_implementation.h
@@ -1,15 +1,6 @@
 #ifndef SIMDJSON_ARM64_DOM_PARSER_IMPLEMENTATION_H
 #define SIMDJSON_ARM64_DOM_PARSER_IMPLEMENTATION_H
 
-#include "simdjson.h"
-#include "isadetection.h"
-
-namespace simdjson {
-namespace arm64 {
-
 #include "generic/dom_parser_implementation.h"
-
-} // namespace arm64
-} // namespace simdjson
 
 #endif // SIMDJSON_ARM64_DOM_PARSER_IMPLEMENTATION_H