--- a/src/arm64/stringparsing.h
+++ b/src/arm64/stringparsing.h
@@ -4,7 +4,6 @@
 #include "simdjson.h"
 #include "jsoncharutils.h"
 #include "arm64/simd.h"
-#include "arm64/intrinsics.h"
 #include "arm64/bitmanipulation.h"
 
 namespace simdjson {