--- a/src/fallback/begin_implementation.h
+++ b/src/fallback/begin_implementation.h
@@ -1,0 +1,4 @@
+#include "simdjson.h"
+#include "fallback/implementation.h"
+
+#define SIMDJSON_IMPLEMENTATION fallback