--- a/src/fallback/begin_implementation.h
+++ b/src/fallback/begin_implementation.h
@@ -2,3 +2,5 @@
 #include "fallback/implementation.h"
 
 #define SIMDJSON_IMPLEMENTATION fallback
+
+#include "fallback/bitmanipulation.h"