--- a/src/fallback/begin_implementation.h
+++ b/src/fallback/begin_implementation.h
@@ -1,6 +1,3 @@
-#include "simdjson.h"
-#include "fallback/implementation.h"
-
 #define SIMDJSON_IMPLEMENTATION fallback
-
+#include "fallback/implementation.h"
 #include "fallback/bitmanipulation.h"