--- a/src/fallback/dom_parser_implementation.cpp
+++ b/src/fallback/dom_parser_implementation.cpp
@@ -4,12 +4,11 @@
 //
 // Stage 1
 //
+#include "generic/stage1/find_next_document_index.h"
+
 namespace simdjson {
 namespace SIMDJSON_IMPLEMENTATION {
-
 namespace stage1 {
-
-#include "generic/stage1/find_next_document_index.h"
 
 class structural_scanner {
 public:
@@ -180,7 +179,6 @@ private:
 
 } // namespace stage1
 
-
 WARN_UNUSED error_code dom_parser_implementation::stage1(const uint8_t *_buf, size_t _len, bool partial) noexcept {
   this->buf = _buf;
   this->len = _len;
@@ -316,14 +314,13 @@ WARN_UNUSED bool implementation::validate_utf8(const char *buf, size_t len) cons
 //
 #include "fallback/stringparsing.h"
 #include "fallback/numberparsing.h"
-
-namespace simdjson {
-namespace SIMDJSON_IMPLEMENTATION {
-
 #include "generic/stage2/logger.h"
 #include "generic/stage2/atomparsing.h"
 #include "generic/stage2/structural_iterator.h"
 #include "generic/stage2/structural_parser.h"
+
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
 
 WARN_UNUSED error_code dom_parser_implementation::parse(const uint8_t *_buf, size_t _len, dom::document &_doc) noexcept {
   error_code err = stage1(_buf, _len, false);