--- a/src/fallback/dom_parser_implementation.cpp
+++ b/src/fallback/dom_parser_implementation.cpp
@@ -314,9 +314,6 @@ WARN_UNUSED bool implementation::validate_utf8(const char *buf, size_t len) cons
 //
 #include "fallback/stringparsing.h"
 #include "fallback/numberparsing.h"
-#include "generic/stage2/logger.h"
-#include "generic/stage2/atomparsing.h"
-#include "generic/stage2/structural_iterator.h"
 #include "generic/stage2/structural_parser.h"
 
 namespace simdjson {