--- a/src/fallback/dom_parser_implementation.h
+++ b/src/fallback/dom_parser_implementation.h
@@ -1,15 +1,6 @@
 #ifndef SIMDJSON_FALLBACK_DOM_PARSER_IMPLEMENTATION_H
 #define SIMDJSON_FALLBACK_DOM_PARSER_IMPLEMENTATION_H
 
-#include "simdjson.h"
-#include "isadetection.h"
-
-namespace simdjson {
-namespace fallback {
-
 #include "generic/dom_parser_implementation.h"
-
-} // namespace fallback
-} // namespace simdjson
 
 #endif // SIMDJSON_FALLBACK_DOM_PARSER_IMPLEMENTATION_H