--- a/src/fallback/implementation.h
+++ b/src/fallback/implementation.h
@@ -26,7 +26,6 @@ public:
 };
 
 } // namespace fallback
-
 } // namespace simdjson
 
 #endif // SIMDJSON_FALLBACK_IMPLEMENTATION_H