--- a/src/fallback/stringparsing.h
+++ b/src/fallback/stringparsing.h
@@ -2,7 +2,6 @@
 #define SIMDJSON_FALLBACK_STRINGPARSING_H
 
 #include "simdjson.h"
-#include "jsoncharutils.h"
 
 namespace simdjson {
 namespace fallback {