--- a/src/generic/stage1/allocate.h
+++ b/src/generic/stage1/allocate.h
@@ -1,3 +1,5 @@
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
 namespace stage1 {
 namespace allocate {
 
@@ -15,3 +17,5 @@ really_inline error_code set_capacity(internal::dom_parser_implementation &parse
 
 } // namespace allocate
 } // namespace stage1
+} // namespace SIMDJSON_IMPLEMENTATION
+} // namespace simdjson