--- a/src/generic/stage1/buf_block_reader.h
+++ b/src/generic/stage1/buf_block_reader.h
@@ -1,3 +1,6 @@
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
+
 // Walks through a buffer in block-sized increments, loading the last part with spaces
 template<size_t STEP_SIZE>
 struct buf_block_reader {
@@ -81,3 +84,6 @@ template<size_t STEP_SIZE>
 really_inline void buf_block_reader<STEP_SIZE>::advance() {
   idx += STEP_SIZE;
 }
+
+} // namespace SIMDJSON_IMPLEMENTATION
+} // namespace simdjson