--- a/src/generic/stage1/json_structural_indexer.h
+++ b/src/generic/stage1/json_structural_indexer.h
@@ -3,6 +3,12 @@
 // We assume the file in which it is included already includes
 // "simdjson/stage1.h" (this simplifies amalgation)
 
+#include "generic/stage1/buf_block_reader.h"
+#include "generic/stage1/json_string_scanner.h"
+#include "generic/stage1/json_scanner.h"
+#include "generic/stage1/json_minifier.h"
+#include "generic/stage1/find_next_document_index.h"
+
 namespace simdjson {
 namespace SIMDJSON_IMPLEMENTATION {
 namespace stage1 {