--- a/src/generic/stage1/utf8_lookup4_algorithm.h
+++ b/src/generic/stage1/utf8_lookup4_algorithm.h
@@ -1,4 +1,5 @@
-using namespace simd;
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
 
 namespace utf8_validation {
 
@@ -173,6 +174,9 @@ using namespace simd;
     }
 
   }; // struct utf8_checker
-}
+} // namespace utf8_validation
 
 using utf8_validation::utf8_checker;
+
+} // namespace SIMDJSON_IMPLEMENTATION
+} // namespace simdjson