--- a/src/generic/stage1/utf8_validator.h
+++ b/src/generic/stage1/utf8_validator.h
@@ -1,3 +1,5 @@
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
 namespace stage1 {
 /**
  * Validates that the string is actual UTF-8.
@@ -24,3 +26,5 @@ bool generic_validate_utf8(const char * input, size_t length) {
 }
 
 } // namespace stage1
+} // namespace SIMDJSON_IMPLEMENTATION
+} // namespace simdjson