--- a/src/generic/stage1/utf8_validator.h
+++ b/src/generic/stage1/utf8_validator.h
@@ -1,6 +1,7 @@
 namespace simdjson {
 namespace SIMDJSON_IMPLEMENTATION {
 namespace stage1 {
+
 /**
  * Validates that the string is actual UTF-8.
  */