--- a/src/generic/stage1/utf8_zwegner_algorithm.h
+++ b/src/generic/stage1/utf8_zwegner_algorithm.h
@@ -1,3 +1,5 @@
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
 //
 // Detect UTF-8 errors.
 //
@@ -358,3 +360,6 @@ struct utf8_checker {
     return (this->special_case_errors.any_bits_set_anywhere() | this->length_errors) ? simdjson::UTF8_ERROR : simdjson::SUCCESS;
   }
 }; // struct utf8_checker
+
+} // namespace SIMDJSON_IMPLEMENTATION
+} // namespace simdjson