--- a/src/generic/stage2/numberparsing.h
+++ b/src/generic/stage2/numberparsing.h
@@ -1,3 +1,5 @@
+namespace simdjson {
+namespace SIMDJSON_IMPLEMENTATION {
 namespace stage2 {
 namespace numberparsing {
 
@@ -472,3 +474,5 @@ really_inline bool parse_number(const uint8_t *const src, W &writer) {
 
 } // namespace numberparsing
 } // namespace stage2
+} // namespace SIMDJSON_IMPLEMENTATION
+} // namespace simdjson