--- a/src/generic/stage2/numberparsing.h
+++ b/src/generic/stage2/numberparsing.h
@@ -1,4 +1,3 @@
-#include "jsoncharutils.h"
 #include <cmath>
 #include <limits>