The generic/ files are pretty tricky to use right, right now, because you have to have the namespaces and regions just right, and you can't #include anything inside a generic file (otherwise the namespace goes wonky).

This PR:
* Increases the amount of potential inlining: defines the TARGET_REGION exactly once, at the beginning, before compiling each implementation, and untargets after.
* Opens the appropriate namespaces
* Moves many #include statements where they are needed
* Compiles the jsoncharutils functions per architecture (tables are still once per executable) to facilitate cross-architecture inlining of these critical functions
* Added really_inline to several functions in jsoncharutils.h

## Things I Didn't Do

* Prevent Multiple Inclusion: I also thought of a way to make it so you could prevent architecture-specific files from being included more than once, which would pretty much make these files exactly the same as any other file; but it would break amalgamation, so I didn't include it. I'll think that one more for the future.
* Ensure All Inline Stuff Everywhere Is Compiled With Target Architecture: I'm not sure how many functions we use from simdjson.h, but I bet there's a few out there. We could compile each target architecture in a separate .o pretty easily, allowing us to compile even common stuff with architecture enabled ...
* hex_to_u32_nocheck: fascinatingly, this is faster when NOT inlined.

## Performance

skylake (haswell) is running faster in general. This is presumably due to the compiler being able to inline more (I imagine it's shy about optimizing across architecture boundaries ...).

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| update-center                    |   8330 |  45.2 |  42.9 |    5% | 142.1 | 142.1 |    0% |   2888 |   2318 |   19% |      0 |      0 |    0% |  25207 |  25222 |    0% |
| twitterescaped                   |   8787 |  42.8 |  41.3 |    3% | 138.1 | 138.1 |    0% |   2230 |   1911 |   14% |      0 |      0 |    0% |  25095 |  25107 |    0% |
| instruments                      |   3442 |  39.0 |  37.8 |    3% | 143.1 | 143.1 |    0% |    263 |     92 |   65% |      0 |      0 |    0% |  10550 |  10222 |    3% |
| gsoc-2018                        |  51997 |  36.5 |  35.8 |    1% | 101.1 | 101.1 |    0% |  15161 |  14857 |    2% |  29778 |  20698 |   30% | 117065 | 116016 |    0% |
| mesh                             |  11306 |  46.5 |  45.7 |    1% | 174.2 | 174.2 |    0% |    886 |    669 |   24% |      0 |      0 |    0% |  43180 |  42959 |    0% |
| twitter                          |   9867 |  43.6 |  43.2 |    1% | 148.8 | 148.8 |    0% |   1808 |   1734 |    4% |      0 |      0 |    0% |  27194 |  27201 |    0% |
| numbers                          |   2345 |  38.8 |  38.4 |    0% | 144.3 | 144.3 |    0% |    110 |    105 |    4% |      0 |      0 |    0% |   7176 |   7138 |    0% |
| marine_ik                        |  46616 |  54.1 |  53.6 |    0% | 180.9 | 180.9 |    0% |   8827 |   8722 |    1% |  59434 |  78497 |  -32% | 185237 | 186333 |    0% |
| apache_builds                    |   1988 |  37.0 |  36.8 |    0% | 139.3 | 139.3 |    0% |     34 |     24 |   29% |      0 |      0 |    0% |   4893 |   4804 |    1% |
| canada                           |  35172 |  45.3 |  45.1 |    0% | 157.6 | 157.6 |    0% |   5376 |   5313 |    1% |  11213 |  12734 |  -13% | 115376 | 115861 |    0% |
| github_events                    |   1017 |  34.6 |  34.5 |    0% | 129.8 | 129.8 |    0% |      8 |      7 |   12% |      0 |      0 |    0% |   1004 |    959 |    4% |
| mesh.pretty                      |  24646 |  35.9 |  35.8 |    0% | 137.7 | 137.7 |    0% |     43 |     33 |   23% |    281 |      0 |    0% |  69239 |  69306 |    0% |
| citm_catalog                     |  26987 |  34.1 |  34.1 |    0% | 130.7 | 130.7 |    0% |    166 |    160 |    3% |      6 |    141 | -2250% |  71737 |  71547 |    0% |
| random                           |   7976 |  57.7 |  57.7 |    0% | 209.3 | 209.3 |    0% |    975 |   1061 |   -8% |      0 |      0 |    0% |  27743 |  27713 |    0% |