@jkeiser 2020-07-16
Presumably because of the increase in inlining, this actually appears to improve performance (and decrease branch misses) somewhat:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| update-center                    |   8330 |  45.2 |  43.2 |    4% | 142.1 | 142.1 |    0% |   2888 |   2530 |   12% |      0 |      0 |    0% |  25207 |  26943 |   -6% |
| twitterescaped                   |   8787 |  42.8 |  41.0 |    4% | 138.1 | 138.1 |    0% |   2230 |   1895 |   15% |      0 |      0 |    0% |  25095 |  26878 |   -7% |
| instruments                      |   3442 |  39.0 |  37.7 |    3% | 143.1 | 143.1 |    0% |    263 |    100 |   61% |      0 |      0 |    0% |  10550 |  11180 |   -5% |
| mesh                             |  11306 |  46.5 |  45.7 |    1% | 174.2 | 174.2 |    0% |    886 |    764 |   13% |      0 |      0 |    0% |  43180 |  46095 |   -6% |
| twitter                          |   9867 |  43.6 |  42.9 |    1% | 148.8 | 148.8 |    0% |   1808 |   1744 |    3% |      0 |      0 |    0% |  27194 |  29122 |   -7% |
| marine_ik                        |  46616 |  54.1 |  53.3 |    1% | 180.9 | 180.9 |    0% |   8827 |   8665 |    1% |  59434 |  71580 |  -20% | 185237 | 193548 |   -4% |
| gsoc-2018                        |  51997 |  36.5 |  36.0 |    1% | 101.1 | 101.1 |    0% |  15161 |  15368 |   -1% |  29778 |  22360 |   24% | 117065 | 122988 |   -5% |
| github_events                    |   1017 |  34.6 |  34.2 |    1% | 129.8 | 129.8 |    0% |      8 |      6 |   25% |      0 |      0 |    0% |   1004 |   1265 |  -25% |
| apache_builds                    |   1988 |  37.0 |  36.6 |    1% | 139.3 | 139.3 |    0% |     34 |     18 |   47% |      0 |      0 |    0% |   4893 |   5024 |   -2% |
| random                           |   7976 |  57.7 |  57.2 |    0% | 209.3 | 209.3 |    0% |    975 |   1042 |   -6% |      0 |      0 |    0% |  27743 |  29770 |   -7% |
| numbers                          |   2345 |  38.8 |  38.4 |    0% | 144.3 | 144.3 |    0% |    110 |    127 |  -15% |      0 |      0 |    0% |   7176 |   7431 |   -3% |
| mesh.pretty                      |  24646 |  35.9 |  35.6 |    0% | 137.7 | 137.7 |    0% |     43 |     34 |   20% |    281 |      0 |    0% |  69239 |  74171 |   -7% |
| canada                           |  35172 |  45.3 |  44.9 |    0% | 157.6 | 157.6 |    0% |   5376 |   5422 |    0% |  11213 |  11758 |   -4% | 115376 | 122861 |   -6% |
| citm_catalog                     |  26987 |  34.1 |  34.0 |    0% | 130.7 | 130.7 |    0% |    166 |    165 |    0% |      6 |    186 | -3000% |  71737 |  76944 |   -7% |