@jkeiser 2020-07-16
@lemire I found a few functions in jsoncharutils.h that had `inline` but not `really_inline` on them. It turns out we're *faster* when hex_to_u32_nocheck is not inlined. (Changing inline to really_inline makes things a tiny bit slower, but removing inline altogether doesn't change it *at all*.) I haven't followed up to find out why, just yet.