@jkeiser 2020-07-16
> > I found a few functions in jsoncharutils.h that had inline but not really_inline on them, and flipped the ones I could. Fascinatingly, it turns out we're faster when hex_to_u32_nocheck is not inlined.
> 
> Given that it is on a rarely taken path, it is not too surprising, right?

Certainly not the most surprising thing in the world, but it's pretty small, so I didn't anticipate it having any real effect. Makes me wonder what else we can make faster with judicious use of functions.