@jkeiser 2020-07-16
Amalgamated headers were broken because of a recursive include (ouch). Fixed that, we'll see if VS is still grumpy.