@jkeiser 2020-07-16
NOTE: the performance gain could well be the normal stage 2 jitters, as well. The primary purpose of this checkin is to make generic/ files work more like normal ones.