@jkeiser 2020-07-17
Looks like it's in good shape now.

The motivation for me here is, I wanted to move the on demand parser into the generic directory so it could use implementation-specific methods ... but those things tend to override simdjson::simdjson_result<simdjson::haswell::stream::document>, which (a) must be defined outside of simdjson::haswell, and (b) must be aware of simdjson::haswell so that it can reference the class as a template parameter. Making the generic files define namespaces instead of assuming them gives us the freedom to do things like that.