@lemire 2020-07-20
@jkeiser There is a mild conflict with what I just merged into master. I think it is easy to fix, but I would rather you do it, so I don't mess things up.