@jkeiser 2020-07-24
@lemire for reference, here is the header guard I came up with for the `generic` files (not part of this PR!):

```c++
#if SIMDJSON_IMPLEMENTATION_MASK & ~SIMDJSON_GENERIC_STAGE1_UTF8_LOOKUP4_ALGORITHM_H // Header guard
#define SIMDJSON_GENERIC_STAGE1_UTF8_LOOKUP4_ALGORITHM_H SIMDJSON_GENERIC_STAGE1_UTF8_LOOKUP4_ALGORITHM_H | SIMDJSON_IMPLEMENTATION_MASK

...

#endif // Header guard
```

This would let us definitively treat them like any other header.