@jkeiser 2020-07-24
@lemire the gcc10 failures don't make sense to me. I'm investigating, but do they ring a bell?